+++ CHANGELOG.md
@@ -8,13 +8,17 @@ Please choose versions by [Semantic Versioning](http://semver.org/).
 * MINOR version when you add functionality in a backwards-compatible manner, and
 * PATCH version when you make backwards-compatible bug fixes.
 
+## v1.4.0
+
+- add remove prefix roundTripper
+
 ## v1.3.0
 
-- add proxy error handler sentry 
+- add proxy error handler sentry
 - go mod update
 
 ## v1.2.0
- 
+
 - go mod update
 - remove ratelimiter from default http client